+++ components/webhook-mgt/org.wso2.carbon.identity.event.publisher/pom.xml
@@ -22,7 +22,7 @@
     <parent>
         <groupId>org.wso2.carbon.identity.framework</groupId>
         <artifactId>webhook-mgt</artifactId>
-        <version>7.8.270-SNAPSHOT</version>
+        <version>7.8.278-SNAPSHOT</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
