+++ components/webhook-mgt/org.wso2.carbon.identity.event.publisher/src/main/java/org/wso2/carbon/identity/event/publisher/api/service/EventPublisher.java
@@ -49,5 +49,5 @@ public interface EventPublisher {
      * @param eventContext Event Context.
      * @throws EventPublisherException If the event publisher cannot handle the event context.
      */
-    void canHandleEvent(EventContext eventContext) throws EventPublisherException;
+    boolean canHandleEvent(EventContext eventContext) throws EventPublisherException;
 }
