+++ components/webhook-mgt/org.wso2.carbon.identity.event.publisher/src/main/java/org/wso2/carbon/identity/event/publisher/api/service/EventPublisherService.java
@@ -41,5 +41,5 @@ public interface EventPublisherService {
      *
      * @param eventContext Event Context.
      */
-    void canHandleEvent(EventContext eventContext) throws EventPublisherException;
+    boolean canHandleEvent(EventContext eventContext) throws EventPublisherException;
 }
